backend/requirements.txt
Django==5.0.3
djangorestframework
django-cors-headers